+++ CHANGELOG.md
@@ -2,12 +2,15 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
+
+## [0.2.1]
 ### Added
 - Disconnecting handle.
+- Load local gateway setting.
 
 ### Changed
 - Modify `updateEndnodeState` method.
-- Upgrade [kii-gateway-agent](https://github.com/ashramwen/kii-gateway-agent) to `0.2.6`.
+- Upgrade [kii-gateway-agent](https://github.com/ashramwen/kii-gateway-agent) to `0.2.9`.
 
 ## [0.1.7] - 2017-03-13
 ### Added
@@ -15,5 +18,6 @@ All notable changes to this project will be documented in this file.
 - Catch `kii-gateway-agent error`.
 - Upgrade [kii-gateway-agent](https://github.com/ashramwen/kii-gateway-agent) to `0.2.5`.
 
-[Unreleased]: https://github.com/ashramwen/kii-mt7688-agent/compare/v0.1.7...HEAD
+[Unreleased]: https://github.com/ashramwen/kii-mt7688-agent/compare/v0.2.1...HEAD
+[0.2.1]: https://github.com/ashramwen/kii-mt7688-agent/compare/v0.1.7...v0.2.1
 [0.1.7]: https://github.com/ashramwen/kii-mt7688-agent/compare/v0.1.6...v0.1.7